Improve the logging when connections fail
This changed does lower the log output when a connection fails to WARN
as the application can always handle this properly. Also is the failed
state forwarded to the application specific part, which may decide if
this is an actual error.

Also the error code get added to the log message.