[#164] Prevent leaking of NetClient instances
The NetClientImpl used by MqttClientImpl to connect to a server
requires its close() method to be invoked to release some resources
which would otherwise prevent it from being GCed properly.

The MqttClientImpl's disconnect() method has been changed to invoke the
NetClientImpl.close() method instead of just the close() method of the
NetSocketImpl instance that represents the connection to the MQTT
server.

Signed-off-by: Kai Hudalla <sophokles.kh@gmail.com>