[#162] Add support for handling timeouts of expected packets (#165)
The MQTT client now supports setting a timeout value and handlers that
will be invoked when an expected packet from the server in response to a
QoS 1 or 2 PUBLISH packet does not arrive in time.

If a response times out, the corresponding PUBLISH packet is removed
from the outbound queue in order to prevent blocking subsequent PUBLISH
messages being via the client.

Signed-off-by: Kai Hudalla <sophokles.kh@gmail.com>