Rewrite the client ping disconnection on a missing response.
The implementation relies on an IdleStateHandler to close the connection when the server has not sent any response. This does not work when the client manually sends a ping request.

We need to change that, so that any call to the ping method will schedule a timer to disconnect the client and cancel this timer when the ping response is received.

close #183
close #184
close #185